fix how classes initialize
